fixed DayOfWeek converter, added NativeType and DateTimeKInd converters

+++ DTOMaker.Models/DayOfWeekConverter.cs
@@ -7,11 +7,11 @@ namespace DTOMaker.Models;
 /// converter allows <see cref="System.DayOfWeek"/> properties to be internally 
 /// stored and serialized as <see cref="System.Int32"/> values.
 /// </summary>
-public sealed class DayOfWeekConverter : IStructConverter<DayOfWeek, int>
+public sealed class DayOfWeekConverter : IStructConverter<DayOfWeek, byte>
 {
-    public NativeType NativeType => NativeType.Int32;
-    public static DayOfWeek ToCustom(int native) => (DayOfWeek)native;
-    public static int ToNative(DayOfWeek custom) => (int)custom;
-    public static DayOfWeek? ToCustom(int? native) => native.HasValue ? ToCustom(native.Value) : null;
-    public static int? ToNative(DayOfWeek? custom) => custom.HasValue ? ToNative(custom.Value) : null;
+    public NativeType NativeType => NativeType.Byte;
+    public static DayOfWeek ToCustom(byte native) => (DayOfWeek)native;
+    public static byte ToNative(DayOfWeek custom) => (byte)custom;
+    public static DayOfWeek? ToCustom(byte? native) => native.HasValue ? ToCustom(native.Value) : null;
+    public static byte? ToNative(DayOfWeek? custom) => custom.HasValue ? ToNative(custom.Value) : null;
 }
