+++ DTOMaker.SrcGen.JsonNewtonSoft.Tests/ConverterTests_Enum.cs
@@ -16,8 +16,8 @@ namespace DTOMaker.SrcGen.JsonNewtonSoft.Tests
                 [Entity(1)]
                 public interface IMyDTO : IEntityBase
                 {
-                    [Member(1, NativeType.Int32, typeof(DTOMaker.Models.DayOfWeekConverter))] DayOfWeek Field1 { get; set; }
-                    [Member(2, NativeType.Int32, typeof(DTOMaker.Models.DayOfWeekConverter))] DayOfWeek? Field2 { get; set; }
+                    [Member(1, NativeType.Byte, typeof(DTOMaker.Models.DayOfWeekConverter))] DayOfWeek Field1 { get; set; }
+                    [Member(2, NativeType.Byte, typeof(DTOMaker.Models.DayOfWeekConverter))] DayOfWeek? Field2 { get; set; }
                 }
             }
             """;
