+++ DTOMaker.SrcGen.JsonSystemText.Tests/ConverterTests_Enum.cs
@@ -16,8 +16,8 @@ namespace DTOMaker.SrcGen.JsonSystemText.Tests
                 [Entity(1)]
                 public interface IMyDTO : IEntityBase
                 {
-                    [Member(1, NativeType.Int32, typeof(DayOfWeekConverter))] DayOfWeek Field1 { get; set; }
-                    [Member(2, NativeType.Int32, typeof(DayOfWeekConverter))] DayOfWeek? Field2 { get; set; }
+                    [Member(1, NativeType.Byte, typeof(DayOfWeekConverter))] DayOfWeek Field1 { get; set; }
+                    [Member(2, NativeType.Byte, typeof(DayOfWeekConverter))] DayOfWeek? Field2 { get; set; }
                 }
             }
             """;
